# docs/openapi.yaml
openapi: 3.0.3
info:
  title: BuildGraph AI MVP API
  version: 0.2.0
  description: BuildGraph 3주 MVP 공통 API 계약. 상세 기준은 docs/API_CONTRACT.md를 따른다.
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
paths:
...
  /api/users:
    post:
      summary: 회원가입
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupRequest"
      responses:
        "201":
          description: 생성된 사용자
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDto"
  /api/auth/login:
    post:
      summary: 이메일 로그인
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: JWT와 사용자 정보
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /api/auth/me:
    get:
      summary: 현재 사용자 조회
      security:
        - bearerAuth: []
      responses:
        "200":
          description: 현재 사용자
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDto"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    Role:
      type: string
      enum: [USER, ADMIN]
...
    UserDto:
      type: object
      required: [id, email, name, role]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        role:
          $ref: "#/components/schemas/Role"
    SignupRequest:
      type: object
      required: [email, password, name, termsAccepted, marketingAccepted]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
        name:
          type: string
        termsAccepted:
          type: boolean
        marketingAccepted:
          type: boolean
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    AuthResponse:
      type: object
      required: [accessToken, refreshToken, user]
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        user:
          $ref: "#/components/schemas/UserDto"
...